widen-system-uid task groups 1,3,4,5: Akka.Remote int->long system uid
- AddressUid.Uid + AddressUidExtension.Uid(ActorSystem) widened to long;
  default generation unchanged (ThreadLocalRandom int-range, rolling-upgrade safe)
- new akka.remote.use-64bit-system-uids switch (default off) enables full-range
  nonzero 64-bit uid generation via netstandard2.0-safe byte-based RNG
- HandshakeInfo/refuseUid, EndpointRegistry, Endpoint.cs (HopelessAssociation,
  ReliableDeliverySupervisor, GotUid, EndpointWriter/Reader), EndpointManager
  messages (Pass/Quarantined/Quarantine/ResendState) re-typed long/long?
- RemoteWatcher HeartbeatRsp/_addressUids/Quarantine re-typed long
- quarantine API hard re-type: IRemoteActorRefProvider/RemoteActorRefProvider/
  RemoteTransport/Remoting.Quarantine(Address, long?); QuarantinedEvent(Address, long)
- MiscMessageSerializer + AkkaPduCodec 32-bit narrowing casts removed (wire
  already 64-bit: ContainerFormats uint64, WireFormats fixed64)

Akka.Cluster intentionally left broken until the next task group commit.

+++ src/core/Akka.Remote.Tests/AddressUidExtensionSpecs.cs
@@ -23,7 +23,7 @@ namespace Akka.Remote.Tests
         [Fact]
         public void AddressUidExtension_should_always_report_same_value()
         {
-            var values = new ConcurrentBag<int>();
+            var values = new ConcurrentBag<long>();
             var parallelOps = 1000;
             var loop = Parallel.For(0, parallelOps, _ =>
             {
